requirements.txt
apache-airflow~=2.0.2
pyspark~=3.1.1
requests~=2.25.1